rename RelayUtil to ShareUtil
