+++ src/main/java/org/thoughtcrime/securesms/ConversationFragment.java
@@ -17,7 +17,7 @@
 package org.thoughtcrime.securesms;
 
 import static com.b44t.messenger.DcContact.DC_CONTACT_ID_SELF;
-import static org.thoughtcrime.securesms.util.RelayUtil.setForwardingMessageIds;
+import static org.thoughtcrime.securesms.util.ShareUtil.setForwardingMessageIds;
 
 import android.annotation.SuppressLint;
 import android.app.Activity;
@@ -66,7 +66,6 @@ import org.thoughtcrime.securesms.util.StickyHeaderDecoration;
 import org.thoughtcrime.securesms.util.Util;
 import org.thoughtcrime.securesms.util.ViewUtil;
 import org.thoughtcrime.securesms.util.views.ConversationAdaptiveActionsToolbar;
-import org.thoughtcrime.securesms.calls.CallUtil;
 
 import java.util.Collections;
 import java.util.LinkedList;
