+++ src/main/java/org/thoughtcrime/securesms/ConversationListFragment.java
@@ -48,7 +48,7 @@ import org.thoughtcrime.securesms.mms.GlideApp;
 import org.thoughtcrime.securesms.notifications.FcmReceiveService;
 import org.thoughtcrime.securesms.permissions.Permissions;
 import org.thoughtcrime.securesms.util.Prefs;
-import org.thoughtcrime.securesms.util.RelayUtil;
+import org.thoughtcrime.securesms.util.ShareUtil;
 import org.thoughtcrime.securesms.util.Util;
 import org.thoughtcrime.securesms.util.ViewUtil;
 
@@ -264,7 +264,7 @@ public class ConversationListFragment extends BaseConversationListFragment
     int listflags = 0;
     if (archive) {
       listflags |= DcContext.DC_GCL_ARCHIVED_ONLY;
-    } else if (RelayUtil.isRelayingMessageContent(getActivity())) {
+    } else if (ShareUtil.isRelayingMessageContent(getActivity())) {
       listflags |= DcContext.DC_GCL_FOR_FORWARDING;
     } else {
       listflags |= DcContext.DC_GCL_ADD_ALLDONE_HINT;
